backend/requirements.txt
# https://www.python.org/dev/peps/pep-0440/#compatible-release
Django~=3.1
psycopg2~=2.8

gunicorn~=19.9
django-rest-framework~=0.1.0
djangorestframework~=3.12.1
drf-yasg~=1.17.1
